Extract operator fusion traits and impls into submodule
Split up the `optimize` module which was getting quite large by extracting a
submodule for graph fusion traits and impls.
